+++ packages/core/src/view/geometry/Shape.ts
@@ -32,10 +32,10 @@ import AbstractCanvas2D from '../canvas/AbstractCanvas2D';
 import SvgCanvas2D from '../canvas/SvgCanvas2D';
 import InternalEvent from '../event/InternalEvent';
 import Client from '../../Client';
-import CellState from '../cell/CellState';
-import StencilShape from './node/StencilShape';
-import CellOverlay from '../cell/CellOverlay';
-import ImageBox from '../image/ImageBox';
+import type CellState from '../cell/CellState';
+import type StencilShape from './node/StencilShape';
+import type CellOverlay from '../cell/CellOverlay';
+import type ImageBox from '../image/ImageBox';
 
 import type {
   ArrowValue,