+++ packages/ts-example/vite.config.js
@@ -27,7 +27,7 @@ export default defineConfig(({ mode }) => {
           },
         },
       },
-      chunkSizeWarningLimit: 468, // @maxgraph/core
+      chunkSizeWarningLimit: 462, // @maxgraph/core
     },
   };
 });